# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List notes
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getNote
      tags: [notes]
      summary: Get a note
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /resources:
    get:
      operationId: listResources
      tags: [resources]
      summary: List resources
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resource"
    post:
      operationId: createResource
      tags: [resources]
      summary: Create a resource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
  /resources/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getResource
      tags: [resources]
      summary: Get a resource
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateResource
      tags: [resources]
      summary: Update a resource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourceUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteResource
      tags: [resources]
      summary: Delete a resource
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Note:
      type: object
      required: [id, userId, title, content, tags, isMarkdown, isPinned, createdAt, updatedAt]
      properties:
        id: { type: string }
        userId: { type: string }
        title: { type: string }
        content: { type: string }
        tags:
          type: array
          items: { type: string }
        isMarkdown: { type: boolean }
        isPinned: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    NoteInput:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1 }
        content: { type: string }
        tags:
          type: array
          items: { type: string }
        isMarkdown: { type: boolean }
    NoteUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        content: { type: string }
        tags:
          type: array
          items: { type: string }
        isMarkdown: { type: boolean }
        isPinned: { type: boolean }
...
    Resource:
      type: object
      required: [id, userId, title, category, isProcessed, createdAt, updatedAt]
      properties:
        id: { type: string }
        userId: { type: string }
        title: { type: string }
        category: { type: string }
        description: { type: ["string", "null"] }
        fileName: { type: ["string", "null"] }
        isProcessed: { type: boolean }
        aiSummary: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ResourceInput:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1 }
        category: { type: string }
        description: { type: string }
        fileName: { type: string }
    ResourceUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        category: { type: string }
        description: { type: string }
        isProcessed: { type: boolean }
        aiSummary: { type: string }
...