# lib/api-zod/src/generated/types/noteInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NoteInput {
  /** @minLength 1 */
  title: string;
  content?: string;
  tags?: string[];
  isMarkdown?: boolean;
}