# lib/api-zod/src/generated/types/resource.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Resource {
  id: string;
  userId: string;
  title: string;
  category: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  fileName?: string | null;
  isProcessed: boolean;
  /** @nullable */
  aiSummary?: string | null;
  createdAt: Date;
  updatedAt: Date;
}