# lib/api-zod/src/generated/types/resourceUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ResourceUpdate {
  /** @minLength 1 */
  title?: string;
  category?: string;
  description?: string;
  isProcessed?: boolean;
  aiSummary?: string;
}